requirements.txt
alembic==1.13.2
asgiref==3.8.1
asyncio==3.4.3
black==24.8.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.1
